Made new BossFight scene, added a boss with a hammer, and added the player and a grid into the scene.
I put scripts into the boss enemy and copied all I needed from bigEnemyScript to bossEnemyScript, and left some things commented out so I knew where I'd need to change behaviors to start attacks. I also have a reference to the bossEnemyAttacksScript in the bossEnemyScript so we can call the methods if we need to.