im fucking with A* Pathfinding inside the LoadNewRoom method in LevelManagerScript but I'm ass.
If you play the game and then manually scan with the grid, it will detect the obstacles and work good. But it doesn't scan after the play button is pressed, it usually get's scanned manually when managing the scene. So i need to find a way to scan it again in the code once the room gets loaded.